[util] do not read hostname from `datadog.conf`
To be able to read the `datadog.conf` Datadog Agent configuration file,
the user must have read privilege to the file (default to `0640
dd-agent:root`).

It is also useless: if a `datadog.conf` file is present on the host, one
can use `statsd` to have metrics, events... sent with the appropriate
hostname.