Check if there are packets to send before sending an empty UDP packet
This avoid cases where the buffer is initialised but then doesn't receive any
packets to send before it's flushed. That can either be because of conditional
sending of metrics or because a sample rate is used which means no packets has
been generated this time around. In those cases an empty UDP packet is still
sent when the buffer is flushed.

This fixes that by checking if there are any packets to send before attempting
to send them off.