Enforce the default 'normal' event priority client side
The CLI help output stats that the default `priority` value used when
posting an event with `dog event post` is `normal`. That value isn't
enforced on the client side, and it seems leaving it to `None` has an
unexpected side effect: the API returns an HTTP 200 response along with
an link to the indexed event, which always returns an HTTP 404.

Example:

```console

$ dog --raw event post --alert_type 'info' --no_host 'test by balthazar' 'checking for any pebcak' | jq .
{
  "status": "ok",
  "event": {
    "id": XXXX,
    "title": "test by balthazar",
    "text": "checking for any pebcak",
    "date_happened": 1578645516,
    "handle": null,
    "priority": null,
    "related_event_id": null,
    "tags": [],
    "url": "https://app.datadoghq.com/event/event?id=1234"
  }
}
```

That event url always returns an HTTP 404 when opened. We can see
`"priority": null` in the response payload.

If we however explicitly set the `--priority` flag to `normal` in the
command-line, we can see `"priority": "normal"` in the JSON output:

```console
$ dog --raw event post --priority normal --alert_type 'info' --no_host 'test by balthazar' 'checking for any pebcak' | jq .
{
  "status": "ok",
  "event": {
    "id": YYYYY,
    "title": "test by balthazar",
    "text": "checking for any pebcak",
    "date_happened": 1578645659,
    "handle": null,
    "priority": "normal",
    "related_event_id": null,
    "tags": [],
    "url": "https://app.datadoghq.com/event/event?id=ABCD"
  }
}
```

Visiting the event URL now returns an HTTP 200, and I can find the event
by searching in the event stream.

Additional API validation will of course have to be added server-side
but this is at least a start.