Add dogshell/dogshellwrap entry points (#305)
Currently, the 'dog' entrypoint conflicts with a binary of the same name
from sheepdog, a package available in major linux distributions. In
packaging, we've renamed the dog entrypoint from this package to
dogshell to match the source module's name and to avoid conflicts. I
think it makes sense to also add it as an entrypoint here for
consistency. I've marked it as Pending Deprecation, so that in the
future we can transition users to dogshell and retire the dog name.