[github] Fix precedence order in CODEOWNERS (#640)
Since the order of items in the CODEOWNERS is important (last matching
line takes precedence), having the `*` line near the bottom of the file
negates any specific patterns above it. By moving this pattern up to the
top of the file, we now ensure that the precedence of ownership is
correctly evaluated.