[statsd] Fix buffer operation thread-safety (#642)
* [statsd] Improve exception when calling buffer ops out-of-order

When a `close_buffer` was called before any `open_buffer`, we would raise
an unexpected exception. This change ensures that we have a consistent
error shown to the user when that happens.

* [statsd] Fix buffer operation thread-safety

Old code did not have locking around its `open_buffer()` and
`close_buffer()` operations which made those two methods not
thread-safe. This change fixes this and aligns this module with the
documentation.