[statsd] Enable buffering by default for statsd
This change turns all sync handling of statsd data into buffered handling
with a thread that forces flushes at a specified interval (default:
300ms, configurable via `flush_interval` kwarg). Because we now do this
buffering by default, the functions `open_buffer()` and `close_buffer()`
are effectively low-level operations and have been deprecated.

To help backwards compatibility where a context manager (which flushes
automatically on exit of the context) is not used, we also now have a
`flush()` function that can be used to force the sending of the metrics
to the server and a `disable_buffering` kwarg in the statsd constructor.