[statsd] Ensure a reasonable minimum send buffer for sockets
On some platforms (e.g. MacOS), the sned/receive socket buffers are
extremely small and are liable to cause issues with sending of the data.
This commit ensures that we have at least 32kb send buffer for our
UDP/UDS sockets so that we encounter less of those errors/problems.