[statsd] Ensure that max fragemtnation flush events are multithreaded-safe
Previously, when a sync flush happened, another thread could have added
data to the end of the buffer, causing the buffer to overflow and the
data to be discarded. By wrapping the fragmentation flush check in the
same lock, we now ensure that the packet won't increase even further
before we send it.