[statsd] Improve tag normalization speed (#672)
* [tests] Add ability to print profiling stats in statsd benchmark test

This change allows us to exactly see where statsd is spending the
majority of its time when processing metrics. The profiling can be
activated via `BENCHMARK_PROFILING` variable.

* [statsd] Improve performance of tag normalization

Previously, we used a raw string in `re.sub` which is in the hot-path
and needed compilation every time so now we use the precompiled version
of the regex to do the substitution, yielding a much better performance.