[statsd] Enable buffering by default for statsd (#670)
* [statsd] Enable buffering by default for statsd

This change turns all sync handling of statsd data into buffered handling
with a thread that forces flushes at a specified interval (default:
300ms, configurable via `flush_interval` kwarg). Because we now do this
buffering by default, the functions `open_buffer()` and `close_buffer()`
are effectively low-level operations and have been deprecated.

To help backwards compatibility where a context manager (which flushes
automatically on exit of the context) is not used, we also now have a
`flush()` function that can be used to force the sending of the metrics
to the server and a `disable_buffering` kwarg in the statsd constructor.

* [statsd] Ensure a reasonable minimum send buffer for sockets

On some platforms (e.g. MacOS), the sned/receive socket buffers are
extremely small and are liable to cause issues with sending of the data.
This commit ensures that we have at least 32kb send buffer for our
UDP/UDS sockets so that we encounter less of those errors/problems.

* [statsd] Ensure that max fragemtnation flush events are multithreaded-safe

Previously, when a sync flush happened, another thread could have added
data to the end of the buffer, causing the buffer to overflow and the
data to be discarded. By wrapping the fragmentation flush check in the
same lock, we now ensure that the packet won't increase even further
before we send it.

* [statsd] Fix formatting/doc nits from code review

There were some spelling and formatting changes that needed to be done
as part of review feedback.