[docs] Fix README docs lang tab query param links (#688)
* [docs] Update README links lang param

`tab`/`lang` language selection query params in doc links has been
changed to `code-lang` so this commit fixes those occurrences
in the README.

* [docs] Fix redirects in README that was losing the lang query param link

ome old links are now redirects to newer paths and losing the lang query
param in the process. We now use the resolved links to preserve the parameter.