[statsd] Disable statsd buffering by default (#692)
* [statsd] Disable buffering by default

Due to impact on users of other clients where buffering was enabled by
default, especially in environments and frameworks where `fork()` is
used we will for the time being disable buffering by default until a
decision is made on the path forward. Buffering can still be turned on
with `disable_buffering = False` flag it's just that for now it isi
defaulted to `True`.

* [statsd] Update notes about buffering in statsd

Since we are at least for the time being disabling buffering by default, the
docs here are being updated to match the changes applied.

* [tests] Ensure that context maanger works in non-buffered envs

With a non-buffered environment, our context manager may not work
properly. This change ensures that we propely test this behavior.