[statsd] Enable toggling of `disable_buffer` flag at runtime
Since `initialize` uses the module-level DogStatsd instance that is only
initialized once, we cannot rely on new instantiation for changes to the
`disable_buffering` flag. This change allows the consumers to call
`stats.disable_buffering = <bool>` with runtime changes being applied to
the module-level instance. By doing this, we can in future changes allow
`initialize()` to change this flag if meeded.