[statsd] Add ability to toggle `statsd.disable_buffering` state during runtime (#700)
* [statsd] Enable toggling of `disable_buffer` flag at runtime

Since `initialize` uses the module-level DogStatsd instance that is only
initialized once, we cannot rely on new instantiation for changes to the
`disable_buffering` flag. This change allows the consumers to call
`stats.disable_buffering = <bool>` with runtime changes being applied to
the module-level instance. By doing this, we can in future changes allow
`initialize()` to change this flag if meeded.

* [init] Add `statsd_disable_buffering` arg to `datadog.initialize()`

Now that the DogStatsd module can handle live toggling of buffering
switches, we expose this argument in the `datadog.initialize()`. While
not terribly useful currently, it will be very valuable if/when
buffering is enabled on the DogStatsd module by default (again).