AMLII-1086 fix stale client_transport tag
initialize reassigns socket_path property on the global statsd
instance, but the client_transport tag was computed in the constructor
and never updated, causing client to misreport transport as udp when a
unix socket was actually used.

This patch makes the tag dynamic and updates it every time the
socket_path property is updated.