Fix potential metric loss when open_buffer is combined with disable_buffering=False (#820)
* Don't reset buffer in open_buffer

When disable_buffering=False, the buffer may contain still unsent
metrics, which are lost when we reset the buffer. When
disable_buffering=True, the reset is a noop anyway.

* Remove invalid todos

Even if disable_buffering defaults to False, the user may still opt to
disable it, in which case the code in these methods still must execute
as it is today.